requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
bson==0.5.10
certifi==2021.10.8
charset-normalizer==2.0.8
curlify==2.2.1
Django==3.2.9
djangorestframework==3.12.4
facebook-business==12.0.1
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pycountry==20.7.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.0
urllib3==1.26.7
yarl==1.7.2
